working multi and div operation and update my Calculator component by adding a child on it
