+++ README.md
@@ -135,12 +135,11 @@ $ npm start
 
 Contributions, issues, and feature requests are welcome!
 
-Feel free to check the [issues page](https://github.com/Bigizi/Math-Magician-React-App).
+Feel free to check the [coming soon]().
 
 <p align="right">(<a href="#readme-top">back to top</a>)</p>
 
 <!-- SUPPORT -->
-h
 ## ⭐️ Show your support <a name="support"></a>
 
 If you like this project, please don't forget to follow us and give it a star.
